subCategoryId로 카테고리내부의 게시물을 가지고 올때 findAll()을 사용하면 안되고 subCategorId로 조회를 해야함으로 생성후 적용

+++ src/main/java/com/devagit/springbootstudy/service/PostService.java
@@ -2,18 +2,20 @@ package com.devagit.springbootstudy.service;
 
 import com.devagit.springbootstudy.domain.post.Post;
 import com.devagit.springbootstudy.exceptionHandler.PostNotFoundException;
+import com.devagit.springbootstudy.exceptionHandler.UserNotFoundException;
 import com.devagit.springbootstudy.repository.post.PostRepository;
 import com.devagit.springbootstudy.view.post.PostListView;
 import com.devagit.springbootstudy.view.post.PostView;
 import org.springframework.stereotype.Service;
 
-import java.text.SimpleDateFormat;
 import java.util.Comparator;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static com.devagit.springbootstudy.exceptionHandler.ErrorCode.POST_NOT_FOUND;
+import static com.devagit.springbootstudy.exceptionHandler.ErrorCode.USER_ID_NOT_THE_SAME;
 
 
 @Service
@@ -24,12 +26,8 @@ public class PostService {
         this.postRepository = postRepository;
     }
 
-    //controller에서 메소드명 동일하게 작성
-    public void addPost(int categoryId, int subCategoryId, String userId, String title, String contents, String source) {
-        long time = System.currentTimeMillis();
-        //datetype으로
-        SimpleDateFormat simpl = new SimpleDateFormat("yyyy:MM:dd hh:mm:ss");
-        String writeDate = simpl.format(time);
+
+    public int addPost(int categoryId, int subCategoryId, String userId, String title, String contents, String source) {
         Post post = Post.builder()      //null값이 들어갈때 명확하게 표현을 해주는게 좋음
                 .categoryId(categoryId)
                 .subCategoryId(subCategoryId)
@@ -37,45 +35,55 @@ public class PostService {
                 .title(title)
                 .contents(contents)
                 .source(source)
-                .writeDate(writeDate)
                 .build();
         postRepository.save(post);
-        //저장하고 아이디 값정도는 보내는게 좋음
+        return post.getId();
     }
 
     public PostView getPost(int id) {
-        Post post = postRepository.findById(id);
+        Post post = Optional.ofNullable(postRepository.findById(id)).orElseThrow(() -> new PostNotFoundException(POST_NOT_FOUND));
         return PostView.from(post);
         //예외처리 꼭하기 명심
     }
 
     public List<PostListView> getPostList(int subCategoryId) {
-        return postRepository.findAll() //subCategoryId로 조회 findAll을 사용하는것이 아니다. (limit를 사용하기),(인덱스를 더공부)
+        return postRepository.findBySubCategoryId(subCategoryId) //subCategoryId로 조회 findAll을 사용하는것이 아니다. (limit를 사용하기),(인덱스를 더공부)
                 .stream()
                 .filter(post -> post.getSubCategoryId() == subCategoryId)
-                .sorted(Comparator.comparing(Post::getWriteDate)) //여기서 하지말고 db에서 꺼낼때 orderBy사용
+                .sorted(Comparator.comparing(Post::getCreateAt)) //여기서 하지말고 db에서 꺼낼때 orderBy사용하는것이 좋음
                 .map(PostListView::from)
+                .limit(20)
                 .collect(Collectors.toList());
     }
 
     public List<PostListView> findPostsByUserId(String userId) {
         List<PostListView> posts = postRepository.findByUserId(userId)
                 .stream()
-                .sorted(Comparator.comparing(Post::getWriteDate))
+                .sorted(Comparator.comparing(Post::getCreateAt))
                 .map(PostListView::from)
                 .collect(Collectors.toList());
-        if (posts.isEmpty()) {
-            throw new PostNotFoundException(POST_NOT_FOUND);
-        } // 리스트로 보낼때는 예외처리를 하지 않고 빈객체를 보낸다
         return posts;
     }
 
-    public void deletePostById(int id) {
-        postRepository.deletePostById(id);
+    public void deletePostById(int id, String userId) {
+        Post post = Optional.ofNullable(postRepository.findById(id)).orElseThrow(() -> new PostNotFoundException(POST_NOT_FOUND));
+        if (userId.equals(post.getUserId())) {
+            postRepository.deletePostById(id);
+        }
     }
 
-    public void updatePost(int id) {
+    public int updatePost(int id, int subCategoryId, String userId, String title, String contents, String source, Date createAt) {
         Post post = postRepository.findById(id);
-        postRepository.save(post); //새로들어온걸 from으로 넣어서 save로 하면댄다 이럴때
+        if (!userId.equals(post.getUserId())) {
+            throw new UserNotFoundException(USER_ID_NOT_THE_SAME);
+        }
+        post.setSubCategoryId(subCategoryId);
+        post.setTitle(title);
+        post.setContents(contents);
+        post.setSource(source);
+        post.setUpdateAt(createAt);
+        postRepository.save(post); //새로들어온걸 from으로 넣어서 save로 하면댄다 이럴때 이해 실패
+        return post.getId();
+
     }
 }
